make resource ready for publish
