+++ src/Resources/RoleResource.php
@@ -3,38 +3,25 @@
 namespace BezhanSalleh\FilamentShield\Resources;
 
 use BezhanSalleh\FilamentShield\Contracts\HasShieldPermissions;
-use BezhanSalleh\FilamentShield\Facades\FilamentShield;
-use BezhanSalleh\FilamentShield\FilamentShieldPlugin;
 use BezhanSalleh\FilamentShield\Forms\ShieldSelectAllToggle;
 use BezhanSalleh\FilamentShield\Resources\RoleResource\Pages;
 use BezhanSalleh\FilamentShield\Support\Utils;
+use BezhanSalleh\FilamentShield\Traits\HasResourceShield;
 use Filament\Facades\Filament;
 use Filament\Forms;
-use Filament\Forms\Components\Component;
 use Filament\Forms\Form;
 use Filament\Resources\Resource;
 use Filament\Tables;
 use Filament\Tables\Table;
 use Illuminate\Contracts\Support\Arrayable;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\HtmlString;
 use Illuminate\Support\Str;
 
 class RoleResource extends Resource implements HasShieldPermissions
 {
-    protected static ?string $recordTitleAttribute = 'name';
+    use HasResourceShield;
 
-    public static function getPermissionPrefixes(): array
-    {
-        return [
-            'view',
-            'view_any',
-            'create',
-            'update',
-            'delete',
-            'delete_any',
-        ];
-    }
+    protected static ?string $recordTitleAttribute = 'name';
 
     public static function form(Form $form): Form
     {
@@ -78,15 +65,7 @@ class RoleResource extends Resource implements HasShieldPermissions
                                 'lg' => 3,
                             ]),
                     ]),
-                Forms\Components\Tabs::make('Permissions')
-                    ->contained()
-                    ->tabs([
-                        static::getTabFormComponentForResources(),
-                        static::getTabFormComponentForPage(),
-                        static::getTabFormComponentForWidget(),
-                        static::getTabFormComponentForCustomPermissions(),
-                    ])
-                    ->columnSpan('full'),
+                static::getShieldFormComponents(),
             ]);
     }
 
@@ -95,11 +74,9 @@ class RoleResource extends Resource implements HasShieldPermissions
         return $table
             ->columns([
                 Tables\Columns\TextColumn::make('name')
-                    // ->badge()
                     ->weight('font-medium')
                     ->label(__('filament-shield::filament-shield.column.name'))
                     ->formatStateUsing(fn ($state): string => Str::headline($state))
-                    // ->color('primary')
                     ->searchable(),
                 Tables\Columns\TextColumn::make('guard_name')
                     ->badge()
@@ -217,205 +194,5 @@ class RoleResource extends Resource implements HasShieldPermissions
     public static function canGloballySearch(): bool
     {
         return Utils::isResourceGloballySearchable() && count(static::getGloballySearchableAttributes()) && static::canViewAny();
-    }
-
-    public static function getResourceEntitiesSchema(): ?array
-    {
-        return collect(FilamentShield::getResources())
-            ->sortKeys()
-            ->map(function ($entity) {
-                $sectionLabel = strval(
-                    static::shield()->hasLocalizedPermissionLabels()
-                    ? FilamentShield::getLocalizedResourceLabel($entity['fqcn'])
-                    : $entity['model']
-                );
-
-                return Forms\Components\Section::make($sectionLabel)
-                    ->description(fn () => new HtmlString('<span style="word-break: break-word;">' . Utils::showModelPath($entity['fqcn']) . '</span>'))
-                    ->compact()
-                    ->schema([
-                        static::getCheckBoxListComponentForResource($entity),
-                    ])
-                    ->columnSpan(static::shield()->getSectionColumnSpan())
-                    ->collapsible();
-            })
-            ->toArray();
-    }
-
-    public static function getResourceTabBadgeCount(): ?int
-    {
-        return collect(FilamentShield::getResources())
-            ->map(fn ($resource) => count(static::getResourcePermissionOptions($resource)))
-            ->sum();
-    }
-
-    public static function getResourcePermissionOptions(array $entity): array
-    {
-        return collect(Utils::getResourcePermissionPrefixes($entity['fqcn']))
-            ->flatMap(function ($permission) use ($entity) {
-                $name = $permission . '_' . $entity['resource'];
-                $label = static::shield()->hasLocalizedPermissionLabels()
-                    ? FilamentShield::getLocalizedResourcePermissionLabel($permission)
-                    : $name;
-
-                return [
-                    $name => $label,
-                ];
-            })
-            ->toArray();
-    }
-
-    public static function setPermissionStateForRecordPermissions(Component $component, string $operation, array $permissions, ?Model $record): void
-    {
-        if (in_array($operation, ['edit', 'view'])) {
-
-            if (blank($record)) {
-                return;
-            }
-            if ($component->isVisible() && count($permissions) > 0) {
-                $component->state(
-                    collect($permissions)
-                        /** @phpstan-ignore-next-line */
-                        ->filter(fn ($value, $key) => $record->checkPermissionTo($key))
-                        ->keys()
-                        ->toArray()
-                );
-            }
-        }
-    }
-
-    public static function getPageOptions(): array
-    {
-        return collect(FilamentShield::getPages())
-            ->flatMap(fn ($page) => [
-                $page['permission'] => static::shield()->hasLocalizedPermissionLabels()
-                    ? FilamentShield::getLocalizedPageLabel($page['class'])
-                    : $page['permission'],
-            ])
-            ->toArray();
-    }
-
-    public static function getWidgetOptions(): array
-    {
-        return collect(FilamentShield::getWidgets())
-            ->flatMap(fn ($widget) => [
-                $widget['permission'] => static::shield()->hasLocalizedPermissionLabels()
-                    ? FilamentShield::getLocalizedWidgetLabel($widget['class'])
-                    : $widget['permission'],
-            ])
-            ->toArray();
-    }
-
-    public static function getCustomPermissionOptions(): ?array
-    {
-        return FilamentShield::getCustomPermissions()
-            ->mapWithKeys(fn ($customPermission) => [
-                $customPermission => static::shield()->hasLocalizedPermissionLabels() ? str($customPermission)->headline()->toString() : $customPermission,
-            ])
-            ->toArray();
-    }
-
-    public static function getTabFormComponentForResources(): Component
-    {
-        return static::shield()->hasSimpleResourcePermissionView()
-            ? static::getTabFormComponentForSimpleResourcePermissionsView()
-            : Forms\Components\Tabs\Tab::make('resources')
-                ->label(__('filament-shield::filament-shield.resources'))
-                ->visible(fn (): bool => (bool) Utils::isResourceEntityEnabled())
-                ->badge(static::getResourceTabBadgeCount())
-                ->schema([
-                    Forms\Components\Grid::make()
-                        ->schema(static::getResourceEntitiesSchema())
-                        ->columns(static::shield()->getGridColumns()),
-                ]);
-    }
-
-    public static function getCheckBoxListComponentForResource(array $entity): Component
-    {
-        $permissionsArray = static::getResourcePermissionOptions($entity);
-
-        return static::getCheckboxListFormComponent($entity['resource'], $permissionsArray, false);
-    }
-
-    public static function getTabFormComponentForPage(): Component
-    {
-        $options = static::getPageOptions();
-        $count = count($options);
-
-        return Forms\Components\Tabs\Tab::make('pages')
-            ->label(__('filament-shield::filament-shield.pages'))
-            ->visible(fn (): bool => (bool) Utils::isPageEntityEnabled() && $count > 0)
-            ->badge($count)
-            ->schema([
-                static::getCheckboxListFormComponent('pages_tab', $options),
-            ]);
-    }
-
-    public static function getTabFormComponentForWidget(): Component
-    {
-        $options = static::getWidgetOptions();
-        $count = count($options);
-
-        return Forms\Components\Tabs\Tab::make('widgets')
-            ->label(__('filament-shield::filament-shield.widgets'))
-            ->visible(fn (): bool => (bool) Utils::isWidgetEntityEnabled() && $count > 0)
-            ->badge($count)
-            ->schema([
-                static::getCheckboxListFormComponent('widgets_tab', $options),
-            ]);
-    }
-
-    public static function getTabFormComponentForCustomPermissions(): Component
-    {
-        $options = static::getCustomPermissionOptions();
-        $count = count($options);
-
-        return Forms\Components\Tabs\Tab::make('custom')
-            ->label(__('filament-shield::filament-shield.custom'))
-            ->visible(fn (): bool => (bool) Utils::isCustomPermissionEntityEnabled() && $count > 0)
-            ->badge($count)
-            ->schema([
-                static::getCheckboxListFormComponent('custom_permissions', $options),
-            ]);
-    }
-
-    public static function getTabFormComponentForSimpleResourcePermissionsView(): Component
-    {
-        $options = FilamentShield::getAllResourcePermissions();
-        $count = count($options);
-
-        return Forms\Components\Tabs\Tab::make('resources')
-            ->label(__('filament-shield::filament-shield.resources'))
-            ->visible(fn (): bool => (bool) Utils::isResourceEntityEnabled() && $count > 0)
-            ->badge($count)
-            ->schema([
-                static::getCheckboxListFormComponent('resources_tab', $options),
-            ]);
-    }
-
-    public static function getCheckboxListFormComponent(string $name, array $options, bool $searchable = true): Component
-    {
-        return Forms\Components\CheckboxList::make($name)
-            ->label('')
-            ->options(fn (): array => $options)
-            ->searchable($searchable)
-            ->afterStateHydrated(
-                fn (Component $component, string $operation, ?Model $record) => static::setPermissionStateForRecordPermissions(
-                    component: $component,
-                    operation: $operation,
-                    permissions: $options,
-                    record: $record
-                )
-            )
-            ->dehydrated(fn ($state) => ! blank($state))
-            ->bulkToggleable()
-            ->gridDirection('row')
-            ->columns(static::shield()->getCheckboxListColumns())
-            ->columnSpan(static::shield()->getCheckboxListColumnSpan());
-    }
-
-    public static function shield(): FilamentShieldPlugin
-    {
-        return FilamentShieldPlugin::get();
     }
 }
