+++ src/Traits/HasPanelShield.php
@@ -11,13 +11,15 @@ trait HasPanelShield
 {
     public static function bootHasPanelShield()
     {
-        if (Utils::isPanelUserRoleEnabled()) {
+        if (! app()->runningInConsole()) {
+            if (Utils::isPanelUserRoleEnabled()) {
 
-            Utils::createPanelUserRole();
+                Utils::createPanelUserRole();
 
-            static::created(fn ($user) => $user->assignRole(Utils::getPanelUserRoleName()));
+                static::created(fn ($user) => $user->assignRole(Utils::getPanelUserRoleName()));
 
-            static::deleting(fn ($user) => $user->removeRole(Utils::getPanelUserRoleName()));
+                static::deleting(fn ($user) => $user->removeRole(Utils::getPanelUserRoleName()));
+            }
         }
     }
